requirements.txt
python-dotenv
streamlit
altair
huggingface-hub
InstructorEmbedding
sentence-transformers
langchain
unstructured # Document loading
chromadb # Vector storage
openai==0.28
openai # For embeddings
tiktoken  # For embeddings
faiss-cpu
pypdf
pypdf2
pyttsx3
SpeechRecognition
PyAudio